refactor: keyFrame 파일 병합
